fix: translate endpoint session

+++ src/controllers/Account/Session/service.ts
@@ -1,10 +1,14 @@
+import { APP_LANG } from '@config/env'
+import { i18nConfig } from '@config/i18nextConfig'
 import { Session, SessionAttributes } from '@database/entities/Session'
 import { validateUUID } from '@expresso/helpers/Formatter'
 import useValidation from '@expresso/hooks/useValidation'
 import { DtoFindAll } from '@expresso/interfaces/Paginate'
+import { ReqOptions } from '@expresso/interfaces/ReqOptions'
 import ResponseError from '@expresso/modules/Response/ResponseError'
 import { queryFiltered } from '@expresso/modules/TypeORMQuery'
 import { Request } from 'express'
+import { TOptions } from 'i18next'
 import { getRepository } from 'typeorm'
 import sessionSchema from './schema'
 
@@ -22,6 +26,10 @@ class SessionService {
    */
   public static async findAll(req: Request): Promise<DtoPaginate> {
     const sessionRepository = getRepository(Session)
+    const { lang } = req.getQuery()
+
+    const defaultLang = lang ?? APP_LANG
+    const i18nOpt: string | TOptions = { lng: defaultLang }
 
     const query = sessionRepository
       .createQueryBuilder()
@@ -33,24 +41,29 @@ class SessionService {
       .getMany()
     const total = await newQuery.getCount()
 
-    return { message: `${total} data has been received.`, data, total }
+    const message = i18nConfig.t('success.data_received', i18nOpt)
+    return { message: `${total} ${message}`, data, total }
   }
 
   /**
    *
    * @param id
+   * @param options
    * @returns
    */
-  public static async findById(id: string): Promise<Session> {
+  public static async findById(
+    id: string,
+    options?: ReqOptions
+  ): Promise<Session> {
     const sessionRepository = getRepository(Session)
+    const i18nOpt: string | TOptions = { lng: options?.lang }
 
-    const newId = validateUUID(id)
+    const newId = validateUUID(id, { ...options })
     const data = await sessionRepository.findOne({ where: { id: newId } })
 
     if (!data) {
-      throw new ResponseError.NotFound(
-        'session data not found or has been deleted'
-      )
+      const message = i18nConfig.t('errors.not_found', i18nOpt)
+      throw new ResponseError.NotFound(`session ${message}`)
     }
 
     return data
@@ -60,20 +73,22 @@ class SessionService {
    *
    * @param UserId
    * @param token
+   * @param options
    * @returns
    */
   public static async findByUserToken(
     UserId: string,
-    token: string
+    token: string,
+    options?: ReqOptions
   ): Promise<Session> {
     const sessionRepository = getRepository(Session)
+    const i18nOpt: string | TOptions = { lng: options?.lang }
 
     const data = await sessionRepository.findOne({ where: { UserId, token } })
 
     if (!data) {
-      throw new ResponseError.Unauthorized(
-        'the login session has ended, please re-login'
-      )
+      const message = i18nConfig.t('errors.session_ended', i18nOpt)
+      throw new ResponseError.Unauthorized(message)
     }
 
     return data
@@ -127,23 +142,19 @@ class SessionService {
   ): Promise<void> {
     const sessionRepository = getRepository(Session)
 
-    const query = sessionRepository
-      .createQueryBuilder()
-      .delete()
-      .where('UserId = :UserId', { UserId })
-      .where('token = :token', { token })
-
-    await query.execute()
+    // delete record
+    await sessionRepository.delete({ UserId, token })
   }
 
   /**
    *
    * @param id
+   * @param options
    */
-  public static async delete(id: string): Promise<void> {
+  public static async delete(id: string, options?: ReqOptions): Promise<void> {
     const sessionRepository = getRepository(Session)
-    const data = await this.findById(id)
 
+    const data = await this.findById(id, { ...options })
     await sessionRepository.delete(data.id)
   }
 }
